split spec_info.go
