+++ load/spec_info.go
@@ -30,33 +30,6 @@ func newSpecInfo(spec *openapi3.T, path string) *SpecInfo {
 	}
 }
 
-type SpecInfoPair struct {
-	Base     *SpecInfo
-	Revision *SpecInfo
-}
-
-func (specInfoPair *SpecInfoPair) GetBaseVersion() string {
-	if specInfoPair == nil {
-		return "n/a"
-	}
-	return specInfoPair.Base.GetVersion()
-}
-
-func (specInfoPair *SpecInfoPair) GetRevisionVersion() string {
-	if specInfoPair == nil {
-		return "n/a"
-	}
-
-	return specInfoPair.Revision.GetVersion()
-}
-
-func NewSpecInfoPair(specInfo1, specInfo2 *SpecInfo) *SpecInfoPair {
-	return &SpecInfoPair{
-		Base:     specInfo1,
-		Revision: specInfo2,
-	}
-}
-
 func getVersion(spec *openapi3.T) string {
 	if spec == nil || spec.Info == nil {
 		return ""
